updated tests and did some refactoring
