--- core/assets/javascripts/multi-select-starts-with.js
@@ -1,53 +1,0 @@
-import LiteTokenFieldCustomSearch from "./lite-tokenfield-starts-with";
-
-const defaultGetItem = (option) => {
-  return { id: option.value, name: option.text, classes: [] };
-};
-
-const getItems = (element, getItem) => {
-  var items = [];
-  var selected = [];
-  for (var i = 0; i < element.options.length; i++) {
-    var option = element.options.item(i);
-    var item = getItem(option);
-    if (option.selected) {
-      selected.push(item);
-    }
-    items.push(item);
-  }
-  return { items, selected };
-};
-
-const tokenFieldSetup = (element, items, selected) => {
-  var tokenField = new LiteTokenFieldCustomSearch({
-    el: element,
-    items: items,
-    newItems: false,
-    addItemOnBlur: true,
-    filterSetItems: false,
-    addItemsOnPaste: true,
-    minChars: 1,
-    itemName: element.name,
-    setItems: selected,
-    keepItemsOrder: false,
-  });
-
-  tokenField._renderItems();
-  tokenField._html.container.id = element.id;
-  tokenField.fieldInput = "";
-  tokenField.originalItems = items;
-  element.remove();
-  return tokenField;
-};
-
-const progressivelyEnhanceMultipleSelectField = (
-  element,
-  getItem = defaultGetItem
-) => {
-  element.parentElement.classList.add("tokenfield-container");
-  var { items, selected } = getItems(element, getItem);
-  var tokenField = tokenFieldSetup(element, items, selected);
-  return tokenField;
-};
-
-export { progressivelyEnhanceMultipleSelectField };
