+++ core/assets/javascripts/multi-select.js
@@ -1,15 +1,11 @@
 import LiteTokenfield from "./lite-tokenfield";
+import LiteTokenFieldStartsWith from "./lite-tokenfield-starts-with";
 
 const defaultGetItem = (option) => {
-  return { id: option.value, name: option.value, classes: [] };
+  return { id: option.value, name: option.text, classes: [] };
 };
 
-const progressivelyEnhanceMultipleSelectField = (
-  element,
-  getItem = defaultGetItem
-) => {
-  element.parentElement.classList.add("tokenfield-container");
-
+const getItems = (element, getItem) => {
   var items = [];
   var selected = [];
   for (var i = 0; i < element.options.length; i++) {
@@ -20,6 +16,15 @@ const progressivelyEnhanceMultipleSelectField = (
     }
     items.push(item);
   }
+  return { items, selected };
+};
+
+const progressivelyEnhanceMultipleSelectField = (
+  element,
+  getItem = defaultGetItem
+) => {
+  element.parentElement.classList.add("tokenfield-container");
+  var { items, selected } = getItems(element, getItem);
   var tokenField = new LiteTokenfield({
     el: element,
     items: items,
@@ -32,10 +37,39 @@ const progressivelyEnhanceMultipleSelectField = (
     setItems: selected,
     keepItemsOrder: false,
   });
+
   tokenField._renderItems();
   tokenField._html.container.id = element.id;
   element.remove();
   return tokenField;
 };
 
-export { progressivelyEnhanceMultipleSelectField };
+const progressivelyEnhanceMultipleSelectFieldStartsWith = (
+  element,
+  getItem = defaultGetItem
+) => {
+  element.parentElement.classList.add("tokenfield-container");
+  var { items, selected } = getItems(element, getItem);
+  var tokenField = new LiteTokenFieldStartsWith({
+    el: element,
+    items: items,
+    newItems: false,
+    addItemOnBlur: true,
+    filterSetItems: false,
+    addItemsOnPaste: true,
+    minChars: 1,
+    itemName: element.name,
+    setItems: selected,
+    keepItemsOrder: false,
+  });
+
+  tokenField._renderItems();
+  tokenField._html.container.id = element.id;
+  element.remove();
+  return tokenField;
+};
+
+export {
+  progressivelyEnhanceMultipleSelectField,
+  progressivelyEnhanceMultipleSelectFieldStartsWith,
+};
